Выделение контроллера фильма
